Major navigation rework
I have been playing around with different navigation layouts. This proposal has the Diataxis framework at the top level of the navigation. The explanation was easy to combine. The tutorials have been labeled snap, charm and capi. The reference was separated into common and then individual snap, charm and capi folders. Finally how tos are not combined yet. This is becuase we do not have feature parity between each of the deployment types